+++ src/test/java/seedu/address/logic/commands/SecLevelTableCommandResultTest.java
@@ -13,21 +13,21 @@ import org.junit.jupiter.api.Test;
 public class SecLevelTableCommandResultTest {
     @Test
     public void equals() {
-        Map<String, Long> titlesValuesMapping1 = new HashMap<>();
-        titlesValuesMapping1.put("Sec 1", 20L);
-        titlesValuesMapping1.put("Sec 2", 30L);
-        titlesValuesMapping1.put("Sec 3", 40L);
-        titlesValuesMapping1.put("Sec 4", 50L);
-        Map<String, Long> titlesValuesMapping2 = new HashMap<>();
-        titlesValuesMapping2.put("Sec 1", 20L);
-        titlesValuesMapping2.put("Sec 2", 30L);
-        titlesValuesMapping2.put("Sec 3", 40L);
-        titlesValuesMapping2.put("Sec 4", 50L);
-        Map<String, Long> titlesValuesMapping3 = new HashMap<>();
-        titlesValuesMapping3.put("Sec 1", 10L);
-        titlesValuesMapping3.put("Sec 2", 10L);
-        titlesValuesMapping3.put("Sec 3", 20L);
-        titlesValuesMapping3.put("Sec 4", 30L);
+        Map<String, Integer> titlesValuesMapping1 = new HashMap<>();
+        titlesValuesMapping1.put("Sec 1", 20);
+        titlesValuesMapping1.put("Sec 2", 30);
+        titlesValuesMapping1.put("Sec 3", 40);
+        titlesValuesMapping1.put("Sec 4", 50);
+        Map<String, Integer> titlesValuesMapping2 = new HashMap<>();
+        titlesValuesMapping2.put("Sec 1", 20);
+        titlesValuesMapping2.put("Sec 2", 30);
+        titlesValuesMapping2.put("Sec 3", 40);
+        titlesValuesMapping2.put("Sec 4", 50);
+        Map<String, Integer> titlesValuesMapping3 = new HashMap<>();
+        titlesValuesMapping3.put("Sec 1", 10);
+        titlesValuesMapping3.put("Sec 2", 10);
+        titlesValuesMapping3.put("Sec 3", 20);
+        titlesValuesMapping3.put("Sec 4", 30);
         CommandResult commandResult1 = new SecLevelTableCommandResult(titlesValuesMapping1);
         CommandResult commandResult2 = new SecLevelTableCommandResult(titlesValuesMapping2);
         CommandResult commandResult3 = new SecLevelTableCommandResult(titlesValuesMapping3);
@@ -53,21 +53,21 @@ public class SecLevelTableCommandResultTest {
 
     @Test
     public void hashcode() {
-        Map<String, Long> titlesValuesMapping1 = new HashMap<>();
-        titlesValuesMapping1.put("Sec 1", 20L);
-        titlesValuesMapping1.put("Sec 2", 30L);
-        titlesValuesMapping1.put("Sec 3", 40L);
-        titlesValuesMapping1.put("Sec 4", 50L);
-        Map<String, Long> titlesValuesMapping2 = new HashMap<>();
-        titlesValuesMapping2.put("Sec 1", 20L);
-        titlesValuesMapping2.put("Sec 2", 30L);
-        titlesValuesMapping2.put("Sec 3", 40L);
-        titlesValuesMapping2.put("Sec 4", 50L);
-        Map<String, Long> titlesValuesMapping3 = new HashMap<>();
-        titlesValuesMapping3.put("Sec 1", 10L);
-        titlesValuesMapping3.put("Sec 2", 10L);
-        titlesValuesMapping3.put("Sec 3", 20L);
-        titlesValuesMapping3.put("Sec 4", 30L);
+        Map<String, Integer> titlesValuesMapping1 = new HashMap<>();
+        titlesValuesMapping1.put("Sec 1", 20);
+        titlesValuesMapping1.put("Sec 2", 30);
+        titlesValuesMapping1.put("Sec 3", 40);
+        titlesValuesMapping1.put("Sec 4", 50);
+        Map<String, Integer> titlesValuesMapping2 = new HashMap<>();
+        titlesValuesMapping2.put("Sec 1", 20);
+        titlesValuesMapping2.put("Sec 2", 30);
+        titlesValuesMapping2.put("Sec 3", 40);
+        titlesValuesMapping2.put("Sec 4", 50);
+        Map<String, Integer> titlesValuesMapping3 = new HashMap<>();
+        titlesValuesMapping3.put("Sec 1", 10);
+        titlesValuesMapping3.put("Sec 2", 10);
+        titlesValuesMapping3.put("Sec 3", 20);
+        titlesValuesMapping3.put("Sec 4", 30);
         CommandResult commandResult1 = new SecLevelTableCommandResult(titlesValuesMapping1);
         CommandResult commandResult2 = new SecLevelTableCommandResult(titlesValuesMapping2);
         CommandResult commandResult3 = new SecLevelTableCommandResult(titlesValuesMapping3);
@@ -82,11 +82,11 @@ public class SecLevelTableCommandResultTest {
 
     @Test
     public void toStringMethod() {
-        Map<String, Long> titlesValuesMapping = new HashMap<>();
-        titlesValuesMapping.put("Sec 1", 20L);
-        titlesValuesMapping.put("Sec 2", 30L);
-        titlesValuesMapping.put("Sec 3", 40L);
-        titlesValuesMapping.put("Sec 4", 50L);
+        Map<String, Integer> titlesValuesMapping = new HashMap<>();
+        titlesValuesMapping.put("Sec 1", 20);
+        titlesValuesMapping.put("Sec 2", 30);
+        titlesValuesMapping.put("Sec 3", 40);
+        titlesValuesMapping.put("Sec 4", 50);
         CommandResult commandResult = new SecLevelTableCommandResult(titlesValuesMapping);
         String expected = SecLevelTableCommandResult.class.getCanonicalName() + "{feedbackToUser="
                 + commandResult.getFeedbackToUser() + ", showHelp=" + commandResult.isShowHelp()
@@ -97,16 +97,16 @@ public class SecLevelTableCommandResultTest {
 
     @Test
     public void parseMappingCorrectly() {
-        Map<String, Long> titlesValuesMapping = new HashMap<>();
-        titlesValuesMapping.put("Sec 1", 20L);
-        titlesValuesMapping.put("Sec 2", 30L);
-        titlesValuesMapping.put("Sec 3", 40L);
-        titlesValuesMapping.put("Sec 4", 50L);
+        Map<String, Integer> titlesValuesMapping = new HashMap<>();
+        titlesValuesMapping.put("Sec 1", 20);
+        titlesValuesMapping.put("Sec 2", 30);
+        titlesValuesMapping.put("Sec 3", 40);
+        titlesValuesMapping.put("Sec 4", 50);
         SecLevelTableCommandResult commandResult = new SecLevelTableCommandResult(titlesValuesMapping);
 
-        assertEquals(commandResult.getSec1Count(), 20L);
-        assertEquals(commandResult.getSec2Count(), 30L);
-        assertEquals(commandResult.getSec3Count(), 40L);
-        assertEquals(commandResult.getSec4Count(), 50L);
+        assertEquals(commandResult.getSec1Count(), 20);
+        assertEquals(commandResult.getSec2Count(), 30);
+        assertEquals(commandResult.getSec3Count(), 40);
+        assertEquals(commandResult.getSec4Count(), 50);
     }
 }
